Found the bug in the sampler plugin that writes outside of the stack
