+++ Game/GameMain/Physics/CollisionSystem.cpp
@@ -305,10 +305,10 @@ void ecs::systems::GroundCollisionSystem::updateEntity(FilteredEntity& _entityIn
 		on_ground = true;
 
 		// TEMP GROUND HIT SOUND
-		//ecs::events::PlaySound sound_event;
-		//sound_event.audioName = AudioName::COIN_TEST_SOUND;
-		//sound_event.soundFlags = SoundFlags::SF_NONE;
-		//createEvent(sound_event);
+		ecs::events::PlaySound sound_event;
+		sound_event.audioName = AudioName::COIN_TEST_SOUND;
+		sound_event.soundFlags = SoundFlags::SF_NONE;
+		createEvent(sound_event);
 	}
 
 	// Break if entity does not move dynamically.
