+++ Game/GameMain_Test/test.cpp
@@ -1,13 +1,13 @@
 //#include "pch.h"
 #include "gtest/gtest.h"
 
-#include "Physics/GroundCollision.h"
-#include "Physics/MovementLogic.h"
-#include "Physics/ObjectCollision.h"
-#include "Physics/FightingLogic.h"
-#include "Physics/QuadTreeTests.h"
+//#include "Physics/GroundCollision.h"
+//#include "Physics/MovementLogic.h"
+//#include "Physics/ObjectCollision.h"
+//#include "Physics/FightingLogic.h"
+//#include "Physics/QuadTreeTests.h"
 #include "Sound/SoundTests.h"
-#include "AiTests.h"
+//#include "AiTests.h"
 #include <crtdbg.h>
 
 int main(int argc, char** argv)
