+++ Sound/Sound_test/pch.h
@@ -8,3 +8,4 @@
 #include "gtest/gtest.h"
 #include <iostream>
 #include <math.h>
+#include <crtdbg.h>
